pipeline_games/requirements.txt
bs4 
pandas
psycopg2-binary
pylint
pytest
python-dotenv
requests
